+++ Code/Libraries/Drivetrain/examples/Test_Sketch/Test_Sketch.ino
@@ -22,8 +22,8 @@ const byte rightMotorForward = 4;
 const byte rightMotorBackward = 5;
 
 int center = 160; //Where the robot aims when it detects a block. Valid values are 0 - 319.
-int deadZone = 20; //How big the "center" of the robot is. Smaller values will cause robot to wiggle more.
-int power = 160; //How much power for wheel motors. Valid values are 0 - 255.
+byte deadZone = 20; //How big the "center" of the robot is. Smaller values will cause robot to wiggle more.
+byte power = 160; //How much power for wheel motors. Valid values are 0 - 255.
 int stepTimes[3] = {1000, 3000, 1000}; //An array where each element is how much time in milliseconds should be spent at each step of rotation.
 
 Drivetrain *wheels;
